update db, test update, auth update

+++ T2207A-SEM3-API/Models/ClassCourse/ClassCourseGradeResponse.cs
@@ -5,7 +5,7 @@
         public int ClassCourseId { get; set; }
         public string ClassName { get; set; }
         public string CourseName { get; set; }
-        public float FinalScore { get; set; }
+        public double FinalScore { get; set; }
         public bool IsPass { get; set; }
     }
 }
